Keep prompt regression coverage within CI budgets
The prompt-spacing and launch-crash regressions need executable coverage, but placing those tests in existing oversized catchall files tripped the Swift file length guard. Move the zsh shell harness into a small support file and keep shortcut file-store regression coverage in its own focused test file so CI can validate the behavior without accepting more file-length debt.

Constraint: Repository policy forbids local test runs; CI owns XCTest execution.

Rejected: Refresh the file length budget | would bless new debt instead of removing it.

Confidence: high

Scope-risk: narrow

Tested: python3 scripts/swift_file_length_budget.py --budget .github/swift-file-length-budget.tsv; git diff --check; plutil -lint GhosttyTabs.xcodeproj/project.pbxproj

Not-tested: Local XCTest suite per repository testing policy
